refactor: expand out trace schema lib
